Update cloud.py
Check to see if hyperparemeters exist in the hyperparameter configuration file before accessing the hyperparameter dictionary by keys. If the hyperparameter(s) does not exist in the hyperparameter configuration file, the script will fail with "key does not exist in dict." There are use cases for only tuning a set of categorical, continuous, or integer hyperparameters.